requirements.txt
flask==2.0.1
tensorflow==2.6.0
numpy==1.19.5
Pillow==8.3.1
supabase==0.7.1
requests==2.26.0
python-dotenv==0.19.0
pytz==2021.1
openai==0.27.8 